Switched some stores to $state runes
Switched over the store used for tabs to a $state rune that dynamically updates the UI when a tab is added or deleted

Switched the stores for global and active extensions to $state runes.

active_id can currently not be converted to a $state run since it is required for it to be reassigned
