+++ src/components/Editor.svelte
@@ -2,43 +2,21 @@
     import "../styles/editor/cm-styles.css"
 
     import type { Extension } from "@codemirror/state";
-    import type { Unsubscriber } from "svelte/store";
 
     import { onMount, onDestroy } from "svelte";
-    import { EditorState, Transaction } from "@codemirror/state";
+    import { EditorState } from "@codemirror/state";
     import { EditorView} from "codemirror";
 
-    import { active_extensions, global_extensions } from "$lib/cm-extensions";
-    import { active_id } from "$lib/tabs";
+    import { active_extensions, global_extensions } from "$lib/cm-extensions.svelte";
 
     let { tab_id }: { tab_id: string} = $props()
 
     let view: EditorView
     let editor_container: HTMLDivElement
-    let unsubscribe_extension_global: Unsubscriber
-    let unsubscribe_extension_active: Unsubscriber
-    let unsubscribe_active_id: Unsubscriber
 
-    let editor_extensions: Extension[] = []
-
-    function update_extensions(){
-        let active_extension_list: Extension[] = $active_id === tab_id ? ($active_extensions[tab_id] || []) : []
-        editor_extensions = [...$global_extensions, ...active_extension_list]
-    }
+    let editor_extensions: Extension[] = [...(active_extensions[tab_id] || []), ...global_extensions]
 
     onMount(() => {
-        update_extensions()
-
-        unsubscribe_extension_global = global_extensions.subscribe(() => {
-            update_extensions()
-        })
-        unsubscribe_extension_active = active_extensions.subscribe(() => {
-            update_extensions()
-        })
-
-        unsubscribe_active_id = active_id.subscribe(() => {
-            update_extensions()
-        })
 
 
         console.log(tab_id)
@@ -52,8 +30,6 @@
     })
 
     onDestroy(() =>{
-        unsubscribe_extension_active()
-        unsubscribe_extension_global()
         view.destroy()
     })
 </script>
