+++ src/components/TabContent.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
     import "../styles/tab.css"
-    import { tabs, active_id } from "../lib/tabs";
+    import { tabs, active_id } from "../lib/tabs.svelte";
 </script>
 
 <div class="tab-content">
-    {#each $tabs as tab (tab.id)}
+    {#each tabs as tab (tab.id)}
     <div class="tab-component {tab.id === $active_id ? 'active' : ''}">
         <tab.element tab_id={tab.id}></tab.element>
     </div>
