+++ src/routes/+page.svelte
@@ -3,7 +3,7 @@
   import Editor from "../components/Editor.svelte";
   import TabBar from "../components/TabBar.svelte";
   import TabContent from "../components/TabContent.svelte";
-  import { createTab } from "../lib/tabs";
+  import { createTab } from "../lib/tabs.svelte";
   createTab(Editor)
 </script>
